# Add binary GCD