@fjarri 2024-10-07
Not for this PR, but I wonder if we can use binary GCD to calculate the Jacobi symbol too. It should have more performance impact since it's called more often during Lucas test, although probably still not too much.