@fjarri 2024-10-07
Also, again not for this PR, but there's a note in `MillerRabin::test()`:
```
    pub fn test(&self, base: &T) -> Primality {
        // TODO: it may be faster to first check that gcd(base, candidate) == 1,
        // otherwise we can return `Composite` right away.
```
Worth checking out since GCD is available in `crypto-bigint` now.